Moved REST code to application module.
